Remove Travis CI config Create Github Actions config Split tox ini into one per supported distro so they can be run seperately
