--- autotest
@@ -1,1 +1,0 @@
-py.test -v -f test/
