--- tox.ini
@@ -1,16 +1,0 @@
-[tox]
-envlist = py37
-
-[testenv]
-whitelist_externals = docker
-deps = -rrequirements.txt
-commands = docker build -f test/debian_9.Dockerfile -t pytest_pihole:debian_9 .
-           docker build -f test/debian_10.Dockerfile -t pytest_pihole:debian_10 .
-           docker build -f test/centos_7.Dockerfile -t pytest_pihole:centos_7 .
-           docker build -f test/centos_8.Dockerfile -t pytest_pihole:centos_8 .           
-           docker build -f test/fedora_31.Dockerfile -t pytest_pihole:fedora_31 .
-           docker build -f test/fedora_32.Dockerfile -t pytest_pihole:fedora_32 .
-           docker build -f test/ubuntu_16.Dockerfile -t pytest_pihole:ubuntu_16 .
-           docker build -f test/ubuntu_18.Dockerfile -t pytest_pihole:ubuntu_18 .
-           docker build -f test/ubuntu_20.Dockerfile -t pytest_pihole:ubuntu_20 .
-           pytest {posargs:-vv -n auto} -m "not build_stage" ./test/
